requirements.txt
easydict==1.12
numpy==1.26.4
Pillow==10.2.0
PyYAML==6.0.1
scikit_learn==1.4.1.post1
torch==2.2.1
torchvision==0.17.1
tqdm==4.64.1
